Added EAGLECAD files for PCB

+++ firmware/src/app_sdcard_write.c
@@ -22,7 +22,7 @@ void APP_SDCARD_WRITE_Initialize(void){
 
 static bool APP_SDCARD_WIRTE_Write_SDCard(
     const DRV_HANDLE handle,
-    int8_t* const pBuffer,
+    int16_t* const pBuffer,
     const uint16_t bytesToWrite,
     uint16_t*const pNumBytesWrote
 )
